# docs/swagger.yaml
definitions:
...
  requests.CreateTask:
    properties:
      answer:
        maxLength: 100
        type: string
      condition:
        maxLength: 2000
        type: string
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - answer
    - condition
    - title
    type: object
...
  requests.DeleteTask:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
...
  requests.EditTask:
    properties:
      answer:
        maxLength: 100
        type: string
      condition:
        maxLength: 2000
        type: string
      id:
        type: integer
      title:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - answer
    - condition
    - id
    - title
    type: object
...
  responses.CreateTaskResponseDTO:
    properties:
      status:
        type: string
      task:
        $ref: '#/definitions/responses.TaskDTO'
    type: object
...
  responses.DeleteTaskResponseDTO:
    properties:
      status:
        type: string
    type: object
...
  responses.EditTaskResponseDTO:
    properties:
      status:
        type: string
      task:
        $ref: '#/definitions/responses.TaskDTO'
    type: object
  responses.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
...
  responses.GetAllTasksResponseDTO:
    properties:
      tasks:
        items:
          $ref: '#/definitions/responses.TaskDTO'
        type: array
    type: object
...
  responses.GetTaskResponseDTO:
    properties:
      status:
        type: string
      task:
        $ref: '#/definitions/responses.TaskDTO'
    type: object
...
  responses.TaskDTO:
    properties:
      answer:
        type: string
      author_id:
        type: integer
      condition:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
  responses.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/task/all:
    get:
      description: Fetches all tasks created by the authenticated user with pagination
      parameters:
      - description: Pagination offset (default is 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Tasks successfully retrieved
          schema:
            $ref: '#/definitions/responses.GetAllTasksResponseDTO'
        "400":
          description: Invalid token or offset
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: No tasks found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Retrieve all tasks
      tags:
      - Tasks
  /api/task/delete:
    delete:
      consumes:
      - application/json
      description: Deletes a specific task by its ID
      parameters:
      - description: Task deletion data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.DeleteTask'
      produces:
      - application/json
      responses:
        "200":
          description: Task successfully deleted
          schema:
            $ref: '#/definitions/responses.DeleteTaskResponseDTO'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Delete a task
      tags:
      - Tasks
  /api/task/edit:
    put:
      consumes:
      - application/json
      description: Updates an existing task with new data
      parameters:
      - description: Task editing data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.EditTask'
      produces:
      - application/json
      responses:
        "200":
          description: Task successfully updated
          schema:
            $ref: '#/definitions/responses.EditTaskResponseDTO'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Edit a task
      tags:
      - Tasks
  /api/task/get/{id}:
    get:
      description: Fetches a specific task by its ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Task successfully retrieved
          schema:
            $ref: '#/definitions/responses.GetTaskResponseDTO'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Retrieve a task by ID
      tags:
      - Tasks
  /api/task/new:
    post:
      consumes:
      - application/json
      description: Creates a new task and saves it to the database
      parameters:
      - description: Task creation data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/requests.CreateTask'
      produces:
      - application/json
      responses:
        "200":
          description: Task successfully created
          schema:
            $ref: '#/definitions/responses.CreateTaskResponseDTO'
        "400":
          description: Invalid token or JSON
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
        "422":
          description: Validation error
          schema:
            $ref: '#/definitions/responses.ValidationErrorResponse'
        "500":
          description: Database error
          schema:
            $ref: '#/definitions/responses.ErrorResponse'
      summary: Create a new task
      tags:
      - Tasks
...
swagger: "2.0"